added 'add request api method'

+++ models/Service.go
@@ -1,9 +1,9 @@
 package models
 
 type Service struct {
-	Id    int    `column:"id"`
+	Id    int64    `column:"id"`
 	CreatedAt string `column:"created_at"`
 	UpdatedAt string `column:"updated_at"`
-	UserId  int `column:"user_id"`
+	UserId  int64 `column:"user_id"`
 	Data  string `column:"data"`
 }
